refactor(#240):공용 모달 컴포넌트 외부에서 isOpend 변경 가능하도록 수정
